# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ServiceResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Alert": {
				"properties": {
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"spread": {
						"type": "number",
						"format": "double"
					},
					"marketName": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"updatedAt",
					"createdAt",
					"spread",
					"marketName",
					"id"
				],
				"type": "object"
			},
			"ServiceResponse_Alert-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Alert"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlertRequest": {
				"properties": {
					"spread": {
						"type": "number",
						"format": "double"
					},
					"market": {
						"type": "string"
					}
				},
				"required": [
					"spread",
					"market"
				],
				"type": "object"
			},
			"AlertManyResponse": {
				"properties": {
					"alerts": {
						"items": {
							"$ref": "#/components/schemas/Alert"
						},
						"type": "array"
					}
				},
				"required": [
					"alerts"
				],
				"type": "object"
			},
			"ServiceResponse_AlertManyResponse-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AlertManyResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AlertResponse": {
				"properties": {
					"alert": {
						"$ref": "#/components/schemas/Alert"
					}
				},
				"required": [
					"alert"
				],
				"type": "object"
			},
			"ServiceResponse_AlertResponse-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AlertResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "spread-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/spread": {
			"get": {
				"operationId": "GetAllSpreads",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/spread/{market}": {
			"get": {
				"operationId": "GetSpread",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "market",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/healthcheck": {
			"get": {
				"operationId": "Check",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/alerts": {
			"post": {
				"operationId": "CreateAlert",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_Alert-or-null_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AlertRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateAlert",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_Alert-or-null_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AlertRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAlert",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_Alert-or-null_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllAlerts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AlertManyResponse-or-null_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/alerts/{market}": {
			"get": {
				"operationId": "GetAlert",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AlertResponse-or-null_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "market",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/alerts/{market}/check": {
			"get": {
				"operationId": "CheckAlert",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "market",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}